force unicode in Python2
After converting to io.write, the Python codebase was
failing under Python 2 because we were writing strings
where unicode was required. We now have enforced unicode
everywhere. As a side-effect, this also made the dumping
of numpy arrays in the word vectors to fail in Python 2.
This is now fixed as well.