Fix bug and fix doubling up of argument
There was a bug that if a single entity is found in all inputs (documents), then the mapply function would simplify the data structure and cause an error. This has been fixed with SIMPLIFY = FALSE. Added a unit test to verify bug is fixed.

Also, passing the input file path to NERAnnotate is redundant since the input file path is set in the cnlp properties. Removed that string argument, added a check that the file parameter exists in cnlp.